End of day

+++ jajapy/SUDD/tests/test_sudd.py
@@ -7,13 +7,13 @@ def test_symbolic_forwards_timed():
     # Arrange
     tau = np.array([[1, 2, 3],
                     [4, 5, 6],
-                    [7, 8, 9]])
+                    [7, 8, 9]], dtype=np.float64)
 
     phi = np.array([[1, 2, 3],
                     [4, 5, 6],
-                    [7, 8, 9]])
+                    [7, 8, 9]], dtype=np.float64)
 
-    pi = np.array([1, 2, 3])
+    pi = np.array([1, 2, 3], dtype=np.float64)
 
     # Act
     alpha = sudd.forwards_symbolic(phi, tau, pi)
@@ -27,13 +27,13 @@ def test_symbolic_backwards_timed():
     # Arrange
     tau = np.array([[1, 2, 3],
                     [4, 5, 6],
-                    [7, 8, 9]])
+                    [7, 8, 9]], dtype=np.float64)
 
     phi = np.array([[1, 2, 3],
                     [4, 5, 6],
-                    [7, 8, 9]])
+                    [7, 8, 9]], dtype=np.float64)
 
-    pi = np.array([1, 2, 3])
+    pi = np.array([1, 2, 3], dtype=np.float64)
 
     # Act
     beta = sudd.backwards_symbolic(phi, tau, pi)
